Separate shell scripts and miscelaneous other startup files
Misc startub scripts are now added to misc/. Shell scripts are mostly
kept in a shell script folder e.g. bash/ with some other files starting
with the shell name in the root directory like bashrc and bash_profile.
The setup.sh script now asks if you want to change your shell and it
only creates links with the relevant desired shell in the home
directory.

Also, it is important to note that the file now links to the bourne
shell (sh) instead of bash. This way it is considerably more portable.